requirements.txt
opencv-python
mediapipe
pyaudio
numpy
Pillow
moviepy